add all prefix completions to python wrapper
