+++ python/src/converters/__init__.py
@@ -1,6 +1,8 @@
-from .pykeyvi_autowrap_conversion_providers import *
+from .match_iterator_converter import *
+from .match_filter_converter import *
 from autowrap.ConversionProvider import special_converters
 
 
 def register_converters():
     special_converters.append(MatchIteratorPairConverter())
+    special_converters.append(MatchFilterConverter())
